requirements.txt
apache-airflow==2.7.0
mlflow==2.8.1
pandas==2.0.3
scikit-learn==1.3.0
kaggle==1.6.11
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
xgboost==1.7.6
psycopg2-binary==2.9.7
requests==2.31.0
lightgbm==4.1.0
pyarrow==14.0.1
optuna==3.3.0